'touch'

+++ app/src/main/java/com/thomasjbarrerasconsulting/faces/kotlin/facedetector/BitmapScaler.kt
@@ -1,6 +1,9 @@
 package com.thomasjbarrerasconsulting.faces.kotlin.facedetector
 
 import android.graphics.Bitmap
+import android.graphics.Canvas
+import android.graphics.Color
+import android.graphics.Paint
 import kotlin.math.max
 import kotlin.math.min
 
@@ -19,10 +22,10 @@ class BitmapScaler {
                 (fullScaleFactor * bitmap.height).toInt(),
                 true
             )
-            val x = max(0, (scaledBitmap.width - width) / 2)
-            val y = max(0, (scaledBitmap.height - height) / 2)
-            val w = min(width, scaledBitmap.width - x)
-            val h = min(height, scaledBitmap.height - y)
+            val x = max(0.0f,  ((scaledBitmap.width - width) / 2.0f) ).toInt()
+            val y = max(0.0f,  ((scaledBitmap.height - height) / 2.0f)).toInt()
+            val w = min(width.toFloat(), scaledBitmap.width.toFloat() - x).toInt()
+            val h = min(height.toFloat(), scaledBitmap.height.toFloat() - y).toInt()
 
             return Bitmap.createBitmap(scaledBitmap, x, y, w, h)
         }
